requirements.txt
requests>=2.31
pandas>=2.0
numpy>=1.24
scipy>=1.11
yfinance>=0.2.31
Nasdaq-Data-Link>=1.0
openpyxl>=3.1
xlrd>=2.0
jinja2>=3.1
plotly>=5.18
streamlit>=1.30
matplotlib>=3.8
seaborn>=0.13
python-dotenv>=1.0
pytest>=7.4
